refactor: split content query repository
